actualization general

+++ app/configuracion/usuarios/page.tsx
@@ -1,6 +1,6 @@
 "use client"
 import { Nav } from "@/components/configuration"
-import { Button2, ButtonRed, Popup, Table } from "@/components/ui"
+import { Button2, ButtonRed, Popup, Spinner2, Table } from "@/components/ui"
 import { PopupNewUser } from "@/components/user"
 import { IUser } from "@/interfaces"
 import axios from "axios"
@@ -13,6 +13,7 @@ export default function Page() {
   const [popup2, setPopup2] = useState({ view: 'hidden', opacity: 'opacity-0', mouse: false })
   const [user, setUser] = useState<IUser>()
   const [shopLogin, setShopLogin] = useState<any>()
+  const [loading, setLoading] = useState(false)
 
   const getUsers = async () => {
     const res = await axios.get(`${process.env.NEXT_PUBLIC_API_URL}/accounts`)
@@ -36,15 +37,19 @@ export default function Page() {
     <>
     <Popup popup={popup2} setPopup={setPopup2}>
       <p>¿Estas seguro que deseas eliminar el usuario {user?.name}?</p>
-      <div className="flex gap-4">
+      <div className="flex gap-2">
         <ButtonRed action={async () => {
-          await axios.delete(`${process.env.NEXT_PUBLIC_API_URL}/shop-login/${user?._id}`)
-          getUsers()
-          setPopup2({ ...popup2, view: 'flex', opacity: 'opacity-0' })
-          setTimeout(() => {
-            setPopup2({ ...popup2, view: 'hidden', opacity: 'opacity-0' })
-          }, 200)
-        }}>Eliminar usuario</ButtonRed>
+          if (!loading) {
+            setLoading(true)
+            await axios.delete(`${process.env.NEXT_PUBLIC_API_URL}/shop-login/${user?._id}`)
+            getUsers()
+            setPopup2({ ...popup2, view: 'flex', opacity: 'opacity-0' })
+            setTimeout(() => {
+              setPopup2({ ...popup2, view: 'hidden', opacity: 'opacity-0' })
+            }, 200)
+            setLoading(false)
+          }
+        }}>{loading ? <Spinner2 /> : 'Eliminar usuario'}</ButtonRed>
         <button onClick={() => {
           setPopup2({ ...popup2, view: 'flex', opacity: 'opacity-0' })
           setTimeout(() => {
